master/requirements.txt
python-dotenv
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-login
flask-wtf